refactor: Rename api_error.py to errors.py
