+++ kuma/rest/_base.py
@@ -7,7 +7,7 @@ from urllib3 import disable_warnings
 from urllib3.exceptions import InsecureRequestWarning
 
 from kuma._logging import configure_logging
-from kuma.rest.api_error import APIError
+from kuma.rest.errors import APIError
 
 _logger = configure_logging()
 _api_version = "v2.1"
