+++ kuma/rest/errors.py
